Partial Refactoring
