+++ adc.h
@@ -1,0 +1,2 @@
+extern void adc_init(void);
+extern uint8_t adc_sample(uint8_t);
