+++ lcd.h
@@ -1,0 +1,3 @@
+extern void lcd_send_command(uint8_t);
+extern void lcd_send_data(char *);
+extern void lcd_init(void);
