+++ therm.h
@@ -1,0 +1,6 @@
+extern uint8_t therm_reset(void);
+extern void therm_write_bit(uint8_t);
+extern uint8_t therm_read_byte(void);
+extern uint8_t therm_read_bit(void);
+extern void therm_write_byte(uint8_t);
+extern void therm_read_temperature(char *);
