+++ usart.c
@@ -1,0 +1,3 @@
+unsigned char USART_Receive(void);
+void USART_Transmit(unsigned char);
+void USART_Init(unsigned int);
